% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleChartOptions.R
\name{googleChartOptions}
\alias{googleChartOptions}
\title{googleChartOptions}
\usage{
googleChartOptions(gChart, ...)
}
\arguments{
\item{gChart}{An object instantiated using \code{\link{googleChart}} to add configuration options to.}

\item{...}{parameters specific to the JS/google library.}

\item{formatter}{An object created using htmlwidget::JS(""), describing a function taking a google
DataTable as input and applying \href{https://developers.google.com/chart/interactive/docs/reference#formatters}{formatters}, as needed.  See, \href{http://detule.github.io/googleCharts/#formatters}{this example}.}

\item{eventHandlers}{An object created using htmlwidget::JS("").}
}
\value{
googleChart
}
\description{
Add options to a googleChart. All options are completely optional. They are documented individually
in the online documentation of the JS library: see for example \href{https://developers.google.com/chart/interactive/docs/gallery/geochart#configuration-options}{configuration options for GeoChart}.
All parameters are accepted without validation, and are converted to arrays and objects using jsonlite::toJSON.
}
\details{
In addition to the JS library parameters, we make note of the following two parameters specific to the [R] package: formatter,
and clickListener, described in more detail in the Arguments section.
}
\note{
See the 
  \href{http://detule.github.io/googleCharts}{online
  documentation} for additional details and examples.
}

